set page title and description to event title and description; only working for local host currently
